requirements.txt
fastapi
uvicorn[standard]
openai
python-multipart
pydantic
python-dotenv
httpx
numpy
requests
qdrant-client
mem0ai
